+++ src/main/java/com/pekka/moni/transaction/dto/TransactionResponse.java
@@ -3,5 +3,7 @@ package com.pekka.moni.transaction.dto;
 import com.pekka.moni.transaction.Transaction;
 import org.springframework.data.domain.Page;
 
-public record TransactionResponse(Page<Transaction> transactions, Double sum){
+import java.math.BigDecimal;
+
+public record TransactionResponse(Page<Transaction> transactions, BigDecimal sum){
 }
